Revert "Remove template comment and switch spaces to tabs"
This reverts commit 1799f7332f947221d0776a735e0ed3a4bad5365d.

This package is forked from Debian, and it's kept in sync with Debian.
Changes from Debian are imported on a regular basis. For every update,
we review the difference between Kali and Debian, to make sure Kali's
diff is still relevant, to check if Kali's diff needs to be updated, and
to make sure no error slipped in.

This review process must be kept as easy as possible, meaning that the
diff between Kali and Debian must be kept to a minimum. We change only
what needs to be changed, and leave everything else untouched. No
whitespace change, no re-indentation, no cosmetic change, nothing.